fixing flags errors

+++ parsing/gnl/get_next_line_utils.c
@@ -99,7 +99,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	return (sub);
 }
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	int	i;
